refactor(product): 헥사고날 아키텍처 적용 및 기존 모듈 제거
- 헥사고날 아키텍처
spring-gift-wishlist/
└── src/
    └── main/
        └── java/
            └── gift/
                ├── common/                (공통 코드: 예외, 어노테이션, 유틸 등)
                └── product/
                    ├── adapter/           (인프라 계층: 외부와 연결)
                    │   ├── web/           (Primary Adapter: Controller 등)
                    │   └── persistence/   (Secondary Adapter: Repository 구현 등)
                    │
                    ├── application/       (애플리케이션 계층: 비즈니스 워크플로우)
                    │   ├── port/          (포트: 계층 간 추상화 인터페이스)
                    │   │   ├── in/        (Input Port: Use Case 인터페이스)
                    │   │   └── out/       (Output Port: 외부 의존 인터페이스)
                    │   └── usecase/       (Use Case 구현체: 실제 워크플로우/조정)
                    │
                    └── domain/            (도메인 계층: 핵심 비즈니스 로직, 엔티티, 도메인 서비스)

- 기존 애플리케이션 서비스 계층 삭제
- 포트와 어댑터 기반 아키텍처로 리팩터링: AddProductUseCase, DeleteProductUseCase 등 분리
- @Adapter와 UseCase 어노테이션 추가로 어댑터 역할 명시
- SOLID의 ISP 원칙에 따라 컨트롤러가 각 유스케이스 인터페이스를 개별 주입받도록 변경
