+++ src/main/java/gift/common/annotation/UseCase.java
@@ -1,0 +1,12 @@
+package gift.common.annotation;
+
+import org.springframework.stereotype.Component;
+
+import java.lang.annotation.*;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Component
+public @interface UseCase {
+}
