--- src/main/java/gift/domain/product/controller/AdminProductController.java
@@ -1,15 +1,0 @@
-package gift.domain.product.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/admin/products")
-public class AdminProductController {
-    
-    @GetMapping
-    public String productsPage() {
-        return "admin/products";
-    }
-}
