--- src/main/java/gift/domain/product/dto/ProductResponse.java
@@ -1,12 +1,0 @@
-package gift.domain.product.dto;
-
-public record ProductResponse(
-        Long id,
-        String name,
-        int price,
-        String imageUrl
-) {
-    public static ProductResponse of(Long id, String name, int price, String imageUrl) {
-        return new ProductResponse(id, name, price, imageUrl);
-    }
-}
