--- src/main/java/gift/domain/product/model/Product.java
@@ -1,36 +1,0 @@
-package gift.domain.product.model;
-
-public class Product {
-
-    private Long id;
-    private String name;
-    private int price;
-    private String imageUrl;
-
-    private Product(Long id, String name, int price, String imageUrl) {
-        this.id = id;
-        this.name = name;
-        this.price = price;
-        this.imageUrl = imageUrl;
-    }
-
-    public static Product of(Long id, String name, int price, String imageUrl) {
-        return new Product(id, name, price, imageUrl);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public String getImageUrl() {
-        return imageUrl;
-    }
-}
