+++ src/main/java/gift/product/application/port/in/DeleteProductUseCase.java
@@ -1,0 +1,5 @@
+package gift.product.application.port.in;
+
+public interface DeleteProductUseCase {
+    void deleteProduct(Long id);
+}
