+++ src/main/java/gift/product/application/port/in/GetProductUseCase.java
@@ -1,0 +1,7 @@
+package gift.product.application.port.in;
+
+import gift.product.application.port.in.dto.ProductResponse;
+
+public interface GetProductUseCase {
+    ProductResponse getProduct(Long id);
+}
