+++ src/main/java/gift/product/application/port/in/UpdateProductUseCase.java
@@ -1,0 +1,7 @@
+package gift.product.application.port.in;
+
+import gift.product.application.port.in.dto.ProductRequest;
+
+public interface UpdateProductUseCase {
+    void updateProduct(Long id, ProductRequest request);
+}
