+++ src/main/java/gift/product/domain/model/Product.java
@@ -1,0 +1,36 @@
+package gift.product.domain.model;
+
+public class Product {
+
+    private final Long id;
+    private String name;
+    private int price;
+    private String imageUrl;
+
+    public Product(Long id, String name, int price, String imageUrl) {
+        this.id = id;
+        this.name = name;
+        this.price = price;
+        this.imageUrl = imageUrl;
+    }
+
+    public static Product of(Long id, String name, int price, String imageUrl) {
+        return new Product(id, name, price, imageUrl);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+}
