+++ src/test/java/gift/common/exception/GlobalExceptionHandlerTest.java
@@ -1,27 +1,24 @@
 package gift.common.exception;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
+import gift.product.adapter.web.ProductController;
+import gift.product.application.port.in.*;
+import gift.product.application.port.in.dto.ProductRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import org.springframework.test.context.ContextConfiguration;
+import org.springframework.mock.web.MockHttpServletResponse;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.web.bind.annotation.*;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 
-
-@Deprecated
-@WebMvcTest
-@ContextConfiguration(classes = {GlobalExceptionHandlerTest.TestController.class, GlobalExceptionHandler.class})
+@WebMvcTest(ProductController.class)
 class GlobalExceptionHandlerTest {
 
     @Autowired
@@ -30,91 +27,35 @@ class GlobalExceptionHandlerTest {
     @Autowired
     private ObjectMapper objectMapper;
 
-    private record TestRequest(
-        @NotBlank(message = "이름은 필수입니다.")
-        String name,
-        @NotNull(message = "나이는 필수입니다.")
-        Integer age
-    ) {}
-
-    @RestController
-    public static class TestController {
-        @PostMapping("/test/validation")
-        public void validation(@Valid @RequestBody TestRequest request) {}
-
-        @GetMapping("/test/illegal-argument")
-        public void illegalArgument() {
-            throw new IllegalArgumentException("잘못된 인자가 전달되었습니다.");
-        }
-
-        @GetMapping("/test/missing-param")
-        public void missingParam(@RequestParam(required = true) String requiredParam) {}
-        
-        @GetMapping("/test/generic-exception")
-        public void genericException() throws Exception {
-            throw new Exception("예상치 못한 에러입니다.");
-        }
-    }
-
+    @MockitoBean
+    private AddProductUseCase addProductUseCase;
+    @MockitoBean
+    private GetProductUseCase getProductUseCase;
+    @MockitoBean
+    private GetAllProductsUseCase getAllProductsUseCase;
+    @MockitoBean
+    private UpdateProductUseCase updateProductUseCase;
+    @MockitoBean
+    private DeleteProductUseCase deleteProductUseCase;
 
     @Test
-    @DisplayName("MethodArgumentNotValidException 처리 테스트")
-    void handleValidationExceptions() throws Exception {
+    @DisplayName("MethodArgumentNotValidException - 검증 실패")
+    void handleMethodArgumentNotValidException() throws Exception {
         // given
-        TestRequest request = new TestRequest(null, null);
+        ProductRequest invalidRequest = new ProductRequest(null, 0, null);
 
-        // when & then
-        mockMvc.perform(post("/test/validation")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(objectMapper.writeValueAsString(request)))
-            .andExpect(status().isBadRequest())
-            .andExpect(jsonPath("$.errorCode").value("VALIDATION_FAILED"))
-            .andExpect(jsonPath("$.errors.length()").value(2))
-            .andExpect(jsonPath("$.errors[?(@.errorType == 'name')].reason").value("이름은 필수입니다."))
-            .andExpect(jsonPath("$.errors[?(@.errorType == 'age')].reason").value("나이는 필수입니다."));
-    }
+        // when
+        MockHttpServletResponse response = mockMvc.perform(post("/api/products")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(invalidRequest)))
+                .andDo(print())
+                .andReturn()
+                .getResponse();
 
-    @Test
-    @DisplayName("HttpMessageNotReadableException 처리 테스트 (잘못된 JSON)")
-    void handleMalformedJson() throws Exception {
-        // given
-        String malformedJson = "{\"name\":\"test\", \"age\":}";
-
-        // when & then
-        mockMvc.perform(post("/test/validation")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(malformedJson))
-            .andExpect(status().isBadRequest())
-            .andExpect(jsonPath("$.errorCode").value("MALFORMED_JSON"))
-            .andExpect(jsonPath("$.errors[0].reason").value("유효하지 않은 요청 본문 형식입니다"));
-    }
-    
-    @Test
-    @DisplayName("IllegalArgumentException 처리 테스트")
-    void handleIllegalArgumentException() throws Exception {
-        // when & then
-        mockMvc.perform(get("/test/illegal-argument"))
-            .andExpect(status().isBadRequest())
-            .andExpect(jsonPath("$.errorCode").value("INVALID_INPUT"))
-            .andExpect(jsonPath("$.errors[0].reason").value("잘못된 인자가 전달되었습니다."));
-    }
-
-    @Test
-    @DisplayName("MissingServletRequestParameterException 처리 테스트")
-    void handleMissingParams() throws Exception {
-        // when & then
-        mockMvc.perform(get("/test/missing-param"))
-            .andExpect(status().isBadRequest())
-            .andExpect(jsonPath("$.errorCode").value("MISSING_PARAMETER"))
-            .andExpect(jsonPath("$.errors[0].errorType").value("requiredParam"));
-    }
-    
-    @Test
-    @DisplayName("처리되지 않은 모든 예외 처리 테스트")
-    void handleGenericException() throws Exception {
-        // when & then
-        mockMvc.perform(get("/test/generic-exception"))
-            .andExpect(status().isInternalServerError())
-            .andExpect(jsonPath("$.errorCode").value("UNEXPECTED_ERROR"));
+        // then
+        assertThat(response.getStatus()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+        assertThat(response.getContentAsString()).contains("VALIDATION_FAILED");
+        assertThat(response.getContentAsString()).contains("입력하신 정보를 다시 확인해주세요");
+        assertThat(response.getContentAsString()).contains("/api/products");
     }
 }
